evilify dired buffers
It is annoying to mentally switch bindings, even more since in non-evilified
buffers SPC must be replaced by M-m.

So even if in dired the binding ctsr (bépo for hjkl) are not working, I can
still use n and p for next and prev line, which should be all I really need. And
other dired bindings are like the ones in vanilla emacs, so I already know them.